Add a simple OTP generator, based on microphone entropy
This adds a dependency on the spritz ciper library, which allows us to
generate a one time password (OTP) using entropy from the microphone as
seed data.

The task that collects data from the microphone is used to fill a
buffer until the buffer is full of data. The OTP generator task uses code
taken from the Spritz examples to generate 32kb of cipher stream from
this entropy and outputs a final hex encoded hash. The task then
disables itself. If the task is run without sufficient entropy, it
re-runs every 200ms (configurable) until there is enough entropy.

At the moment, all values are wiped once the OTP is printed. This OTP
will be used for verification on the server side when adding a device,
so in the future we will need to not wipe this value until after it is
verified.