Add initial https functionality
This adds very primitive https functionality, with some trials an
tribulations along the way.

Firstly, I located a fork of the ESP8266 Arduino framework that had
rudimentary https support. Next, in order for this to actually be
included and compiled in my platformio code, I had to fork the ESP8266
platformio framework, and adjust the config files to point to this
forked arduino library. Thankfully, platformio segregates things well
enough that this fork defined in the local platformio.ini can be kept
separate from the global install.

After that it was fairly simple to get things working using the
examples provided in the modified framework. I have been able to get
this to work, but only sporadically and only once before it doesn't work
any more.

Work for next time:
Look at the example at https://github.com/esp8266/Arduino/compare/master...earlephilhower:master
that uses WiFiServerSecure. The reason things don't work using
WebServerSecure is likely that this is only meant to handle a single
client. We should implement a modified version of WiFiServerSecure in
order to allow for efficient scheduling with other tasks, and more
reliable message delivery.