Add more reliable web server implementation
This adds a slightly more reliable method of responding to https
requests using the WiFiSecureServer class. However we have to handle
responses manually this way, so we may want to wrap this a little
better to allow for handling responses in a more understandable fashion.

Code is based on example code provided as part of the WiFiSecureServer
implementation in the ESP8266 framework.