Implement parsing of request to verify OTP
This implements a very primitive way of parsing a request from a server
in order to validate the OTP.

At the moment we are simply looking for the first equals sign in the
HTTP request, and then taking everything after that up until the space
to be the OTP. If it matches, we set our serverIP to the IP that sent
the request and respond with a 200. If not, we respond with a 403.

We also implement a little bit of logic to effectively lock the event
loop in one of a few states, so that https requests are processed in a
timely and reliable fashion. If we try and run the request processing
logic at the same time as the task scheduler loop, we often lose
requests. This logic basically will execute the task scheduler until
the entropy pool is full, and will also execute the task scheduler until
the OTP is set. Then it will not execute until the server IP is set,
which basically means it will wait and just process http. Then it will
go back to the scheduler loop once the server is set.

In the future it would be a good idea to implement more robust handling
of the requests from the prospective servers.