Successfully post data to a server using hardcoded ip and host
This successfully will post data to a dummy nginx server and receives
and recognises a 200 response. This should be enough to move on from
this piece of work (because it's getting boring).

Still to do:
* For some reason the server cert validation fails after the first
connection. Not sure why. Need to resolve.
* Also need to see if we can use a custom dns server so that we can
resolve proper certs with proper names eg.
light-pet-server.local.skouf.com. Using IPs is no good.
* Refactor to use the HttpClient object, though I think I tried with
this at some point and the downside is that it blocks?
* Look into session resumption, seems wasteful to kill the http
connection each time.
* Fix issue where a few packets of data are sent before verification of
the OTP